v2.0

+++ api/src/test/resources/schema.sql
@@ -10,7 +10,7 @@ CREATE TABLE IF NOT EXISTS products
     name        VARCHAR(255)                            NOT NULL,
     description VARCHAR(2550)                           NOT NULL,
     image       VARCHAR(2550)                           NOT NULL,
-    price       DECIMAL(10, 2)                          NOT NULL,
+    price       DOUBLE PRECISION                         NOT NULL,
     CONSTRAINT pk_products PRIMARY KEY (id)
 );
